Update code ExtraPractice

+++ Services/Fsel.Course/Fsel.Course.Infrastructure/Common/ExtraPracticeConverter.cs
@@ -301,9 +301,11 @@ namespace Fsel.Course.Infrastructure.Common
                 exercises = extraPractice.Video.VideoTimeCodes.SelectMany(x => x.TimeCodeExercises).Select(x => x.Exercise ?? new Exercise()).ToList();
                 questions = exercises.SelectMany(x => x.ExerciseQuestions).Select(x => x.Question ?? new Question()).ToList();
             }
-
-            if (extraPracticeExercises != null)
+            if (extraPracticeExercises != null && extraPracticeExercises.Count > 0)
             {
+                exercises = extraPracticeExercises.Select(x => x.Exercise ?? new Exercise()).ToList();
+                questions = exercises.SelectMany(x => x.ExerciseQuestions).Select(x => x.Question ?? new Question()).ToList();
+
                 foreach (var item in extraPracticeExercises)
                 {
                     await _extraPracticeExerciseRepository.DeleteAsync(item);
